+++ src/utils/__tests__/findCommentsInRaws.js
@@ -16,7 +16,7 @@ test("Various.", t => {
       a
       {}
       
-      b // comment 2 
+      b // there's a trailing whitespace at the end 
       {
         width: /* comment 3 */ 100px;
         background: url(http://lol);
@@ -31,19 +31,40 @@ test("Various.", t => {
       t.equal(comments[0].raws.startToken, "/**!")
       t.equal(comments[0].raws.endToken, "*/")
       t.equal(comments[0].inlineAfter, false)
+      t.deepEqual(comments[0].source, { start: 7, end: 21 })
       t.equal(comments[1].type, "double-slash")
-      t.equal(comments[1].text, "comment 2")
-      t.equal(comments[1].start, 55)
-      t.equal(comments[1].end, 67)
+      t.equal(comments[1].text, "there's a trailing whitespace at the end")
+      t.deepEqual(comments[1].source, { start: 55, end: 98 })
       t.equal(comments[1].inlineAfter, true)
       t.equal(comments[2].inlineBefore, true)
       t.equal(comments[2].inlineAfter, true)
-      t.equal(comments[2].start, 92)
+      // Here we take into account the fact that postcss-scss transforms
+      // `// comment` -> `/* comment*/`, thus adding 2 symbols to the end
+      t.deepEqual(comments[2].source, { start: 125, end: 139 })
     })
     .catch(logError)
 })
 
-test("", t => {
+test("//", t => {
+  t.plan(4)
+
+  postcss()
+    .process(`
+      //
+    `, { syntax: scss })
+    .then(result => {
+      const css = result.root.source.input.css
+      const comments = findCommentsInRaws(css)
+      
+      t.equal(comments.length, 1)
+      t.equal(comments[0].text, "")
+      t.equal(comments[0].inlineAfter, false)
+      t.deepEqual(comments[0].source, { start: 7, end: 8 })
+    })
+    .catch(logError)
+})
+
+test("// Inline comment, after {.", t => {
   t.plan(3)
 
   postcss()
@@ -63,7 +84,7 @@ test("", t => {
     .catch(logError)
 })
 
-test("", t => {
+test("} // comment", t => {
   t.plan(2)
 
   postcss()
@@ -97,7 +118,7 @@ test("Triple-slash comment", t => {
 })
 
 test("Some fancy comment", t => {
-  t.plan(4)
+  t.plan(5)
 
   postcss()
     .process(`
@@ -114,6 +135,7 @@ test("Some fancy comment", t => {
       t.equal(comments[0].inlineAfter, false, "Inline after something")
       t.equal(comments[0].raws.left, `
        `, "Raws left.")
+      t.deepEqual(comments[0].source, { start: 7, end: 189 })
     })
     .catch(logError)
 })
@@ -139,27 +161,30 @@ test("Another fancy comment", t => {
     .catch(logError)
 })
 
-test("Another fancy comment", t => {
-  t.plan(4)
+test("Comments inside comments", t => {
+  t.plan(7)
 
   postcss()
     .process(`
       /* Text.. /* is that a new comment? */
-      /* And // this? */
       // And /* this */ ?
+      /* And // this? */
     `, { syntax: scss })
     .then(result => {
       const css = result.root.source.input.css
       const comments = findCommentsInRaws(css)
       t.equal(comments.length, 3)
       t.equal(comments[0].text, "Text.. /* is that a new comment?", "text 1")
-      t.equal(comments[1].text, "And // this?", "text 2")
-      t.equal(comments[2].text, "And /* this */ ?", "text 3")
+      t.deepEqual(comments[0].source, { start: 7, end: 44 })
+      t.equal(comments[1].text, "And /* this */ ?", "text 3")
+      t.deepEqual(comments[1].source, { start: 52, end: 70 })
+      t.equal(comments[2].text, "And // this?", "text 2")
+      t.deepEqual(comments[2].source, { start: 80, end: 97 })
     })
     .catch(logError)
 })
 
-test("Another fancy comment", t => {
+test("No comments, but parsing a selector with ().", t => {
   t.plan(1)
 
   postcss()
@@ -176,6 +201,62 @@ test("Another fancy comment", t => {
       const css = result.root.source.input.css
       const comments = findCommentsInRaws(css)
       t.equal(comments.length, 0)
+    })
+    .catch(logError)
+})
+
+test("//-comment, Unix newlines", t => {
+  t.plan(2)
+
+  postcss()
+    .process("\n   // comment \n", { syntax: scss })
+    .then(result => {
+      const css = result.root.source.input.css
+      const comments = findCommentsInRaws(css)
+      t.equal(comments.length, 1)
+      t.deepEqual(comments[0].source, { start: 4, end: 14 })
+    })
+    .catch(logError)
+})
+
+test("CSS comment, Unix newlines", t => {
+  t.plan(2)
+
+  postcss()
+    .process("\n   /* part 1 \n part 2*/ \n", { syntax: scss })
+    .then(result => {
+      const css = result.root.source.input.css
+      const comments = findCommentsInRaws(css)
+      t.equal(comments.length, 1)
+      t.deepEqual(comments[0].source, { start: 4, end: 23 })
+    })
+    .catch(logError)
+})
+
+test("//-comment, Windows-newline", t => {
+  t.plan(2)
+
+  postcss()
+    .process("\r\n   // comment \r\n", { syntax: scss })
+    .then(result => {
+      const css = result.root.source.input.css
+      const comments = findCommentsInRaws(css)
+      t.equal(comments.length, 1)
+      t.deepEqual(comments[0].source, { start: 5, end: 15 })
+    })
+    .catch(logError)
+})
+
+test("CSS comment, Windows newlines", t => {
+  t.plan(2)
+
+  postcss()
+    .process("\r\n   /* part 1 \r\n part 2*/ \r\n", { syntax: scss })
+    .then(result => {
+      const css = result.root.source.input.css
+      const comments = findCommentsInRaws(css)
+      t.equal(comments.length, 1)
+      t.deepEqual(comments[0].source, { start: 5, end: 25 })
     })
     .catch(logError)
 })