+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fixed: `findCommentsInRaws` fail on parsing selectors like `p:not(.not-p)` (applicable to rules `double-slash-comment-inline`, `double-slash-comment-whitespace-inside`, `operator-no-unspaced`).
 - Fixed: 'double-slash-comment-whitespace-inside' false positives on empty comments (e.g. `//`).
+- Fixed: `findCommentsInRaws` giving wrong column number (applicable to rules `double-slash-comment-inline`, `double-slash-comment-whitespace-inside`, `operator-no-unspaced`).
 
 # 1.3.1